做一次短期合并。

+++ src/main/java/com/pojo/Order.java
@@ -48,7 +48,7 @@ public class Order {
     /**
      * 金额
      */
-    private Long total_price;
+    private Double total_price;
     /**
      * 状态
      */
@@ -115,11 +115,11 @@ public class Order {
         this.addr = addr;
     }
 
-    public Long getTotal_price() {
+    public Double getTotal_price() {
         return total_price;
     }
 
-    public void setTotal_price(Long total_price) {
+    public void setTotal_price(Double total_price) {
         this.total_price = total_price;
     }
 }
